상품 장바구니에서 주문하기 구현 중

+++ src/main/java/com/spharosacademy/project/SSGBack/category/dto/output/ProductOfCategory.java
@@ -16,5 +16,5 @@ public class ProductOfCategory {
     private String brand;
     private String mallTxt;
     private String thumbnailUrl;
-    private int price;
+    private float price;
 }
